requirements.txt
Adafruit-Blinka==8.0.2
adafruit-circuitpython-busdevice==5.1.10
adafruit-circuitpython-motor==3.4.2
adafruit-circuitpython-pca9685==3.4.3
adafruit-circuitpython-register==1.9.10
adafruit-circuitpython-servokit==1.3.10
adafruit-circuitpython-typing==1.7.1
adafruit-circuitpython-vl53l0x==3.6.3
Adafruit-PlatformDetect==3.25.0
Adafruit-PureIO==1.1.9
isort==5.10.1
pkg-resources==0.0.0
pyftdi==0.54.0
pyserial==3.5
pyusb==1.2.1
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing-extensions==4.3.0
